==> notebooks/2. Generate Ground Truth Data.ipynb <==
import pandas as pd

df = pd.read_csv("../data/processed_data/all_cities_combined_clean.csv")

df

# change csv to json 
documents = df.to_dict(orient='records')

documents[495:500]

## Create hashid for the text 

import hashlib
from tqdm.auto import tqdm
import json

def generate_document_id(doc):
    combined = f"{doc['city']}-{doc['section']}-{doc['text'][:50]}" # to be sure the data is unqiue 
    hash_object = hashlib.md5(combined.encode())
    hash_hex = hash_object.hexdigest()
    document_id = hash_hex[:8]
    return document_id

for doc in tqdm(documents):
    doc['id'] = generate_document_id(doc)

documents[:6]

# Extract all IDs
ids = [doc["id"] for doc in documents]

# Check duplicates
duplicates = set([i for i in ids if ids.count(i) > 1])

if duplicates:
    print("Duplicated IDs found:", duplicates)
else:
    print("All IDs are unique")


# Save the updated json file 

with open("../data/processed_data/documents-with-ids.json", "wt") as f_out:
    json.dump(documents, f_out, indent=2)

!head ../data/processed_data/documents-with-ids.json

## Genrated the ground_truth dataste 

prompt_template = """
You are a traveller using our travel assistant app.
From the provided travel text, create 5 realistic questions a traveller might ask.  
- Each question must be specific and self-contained.  
- Do not copy long phrases verbatim from the text; paraphrase naturally.  
- Each question must have a short, factual answer derived only from the text.  
- Avoid generic or overly broad questions.

The record: 
City: {city},
Section: {section},
Subsection: {subsection},
Text: {text}

Provide the output in parsable JSON without using code blocks:

{{"questions": ["question1", "question2", ..., "question5"]}}
""".strip()

prompt = prompt_template.format(**documents[0])

print(prompt)

from mistralai import Mistral
from mistralai.models import UserMessage
import os
from dotenv import load_dotenv

# loads variables from .env
load_dotenv()  

api_key = os.getenv("API_KEY")

client = Mistral(api_key = api_key)
response = client.chat.complete(
    model= "ministral-8b-latest",
    messages=[UserMessage(content=prompt)],
)

print(response.choices[0].message.content)

## Generate Questions


def generate_question(doc):
    prompt = prompt_template.format(**doc)
    client = Mistral(api_key = api_key)
    response = client.chat.complete(
        model= 'mistral-large-2411',
        messages=[UserMessage(content=prompt)],
    )    
    json_response = response.choices[0].message.content
    return json_response

results = {}

for doc in tqdm(documents):
    doc_id = doc['id']
    if doc_id in results:
        continue 
    question_raw = generate_question(doc)
    question = json.loads(question_raw)
    results[doc_id]=question['questions']

results['845d3951']

len(results)

doc_index = {}
for d in documents:
    doc_index[d["id"]] = d

final_results = []
for doc_id, qs in results.items():
    city = doc_index[doc_id]['city']  
    for q in qs:
        final_results.append((doc_id, city, q))

final_results[:5]

print(f"The length of the final result is: {len(final_results)}")

df_gt = pd.DataFrame(final_results, columns=['id', 'city', 'question'])

df_gt.head(5)

RESULT_FOLDER = "../data/result"
os.makedirs(RESULT_FOLDER, exist_ok=True)

df_gt.to_csv(f"{RESULT_FOLDER}/groud-truth-retrieval.csv", index=False)
