==> notebooks/3. Minisearch Text Evaluate.ipynb <==
# import os
# os.environ["SSL_CERT_FILE"] = "/mnt/d/Travel Assistant/Musafir/Fortinet_CA_SSL(15).cer"
# os.environ["REQUESTS_CA_BUNDLE"] = "/mnt/d/Travel Assistant/Musafir/Fortinet_CA_SSL(15).cer"

import json
import pandas as pd
from tqdm.auto import tqdm
import os

with open('../data/processed_data/documents-with-ids.json', 'rt') as f_in:
    documents = json.load(f_in)


documents[500]

!wget https://raw.githubusercontent.com/alexeygrigorev/minsearch/main/minsearch.py

import minsearch

df_gt = pd.read_csv('../data/result/groud-truth-retrieval.csv')

df_gt.head()

ground_truth = df_gt.to_dict(orient='records')


ground_truth[:5]

index = minsearch.Index(
    text_fields=['section', 'text'],
    keyword_fields=['id', 'city']
)

index.fit(documents)

# Recall
def recall(relevance_total):
    scores = []
    for line in relevance_total:
        total_relevant = sum(line)
        if total_relevant == 0:
            scores.append(0)
        else:
            retrieved_relevant = sum(line)
            scores.append(retrieved_relevant / total_relevant)
    return sum(scores) / len(scores)

#Hit Rate (HR) or Recall at k
def hit_rate(relevance_total):
    cnt = 0

    for line in relevance_total:
        if True in line:
            cnt = cnt + 1

    return cnt / len(relevance_total)

# Mean Reciprocal Rank (MRR)
def mrr(relevance_total):
    total_score = 0.0

    for line in relevance_total:
        for rank in range(len(line)):
            if line[rank] == True:
                total_score = total_score + 1 / (rank + 1)

    return total_score / len(relevance_total)

def minsearch_search(query):
    boost = {}

    results = index.search(
        query=query,
        filter_dict={},
        boost_dict=boost,
        num_results=10
    )

    return results


def evaluate(ground_truth, search_function):
    relevance_total = []

    for q in tqdm(ground_truth):
        doc_id = q['id']
        results = search_function(q)
        relevance = [d['id'] == doc_id for d in results]
        relevance_total.append(relevance)

    return {
        'Recall': recall(relevance_total),
        'hit_rate': hit_rate(relevance_total),
        'mrr': mrr(relevance_total),
    }
        

evaluate(ground_truth, lambda q: minsearch_search(q['question']))

def minsearch_search_filter(query, city):
    boost = {'text': 3.0, 'section': 0.5}

    results = index.search(
        query=query,
        filter_dict={'city': city},
        boost_dict=boost,
        num_results=5
    )

    return results
    

evaluate(ground_truth, lambda q: minsearch_search_filter(q['question'], q['city']))

### Baseline Minsearch
{'Recall': 0.8552486187845304,

 'hit_rate': 0.8552486187845304,
 
 'mrr': 0.7135870677307133}

## Finding the best paramters


df_validation = df_gt[:150]
df_test = df_gt[150:]

import random

def simple_optimize(param_ranges, objective_function, n_iterations=10):
    best_params = None
    best_score = float('-inf')  # Assuming we're minimizing. Use float('-inf') if maximizing.

    for _ in range(n_iterations):
        # Generate random parameters
        current_params = {}
        for param, (min_val, max_val) in param_ranges.items():
            if isinstance(min_val, int) and isinstance(max_val, int):
                current_params[param] = random.randint(min_val, max_val)
            else:
                current_params[param] = random.uniform(min_val, max_val)
        
        # Evaluate the objective function
        current_score = objective_function(current_params)
        
        # Update best if current is better
        if current_score > best_score:  # Change to > if maximizing
            best_score = current_score
            best_params = current_params
    
    return best_params, best_score

gt_val = df_validation.to_dict(orient='records')

param_ranges = {
 'city': (0.0, 3.0),
 'section': (0.0, 3.0),
 'text': (0.0, 3.0),
}

def minsearch_search_par(query, boost=None):
    
    if boost is None:
        boost = {}

    results = index.search(
        query=query,
        filter_dict={},
        boost_dict=boost,
        num_results=10
    )

    return results

def objective(boost_params):
    def search_function(q):
        return minsearch_search_par(q['question'], boost_params)

    results = evaluate(gt_val, search_function)
    return results['mrr']

simple_optimize(param_ranges, objective, n_iterations=40)

# Apply the boost params in the whole dataset
def minsearh_improved(query):
    
    boost = {
          'city': 2.91,
          'section': 0.17,
          'text': 2.93,
    }

    results = index.search(
        query=query,
        filter_dict={},
        boost_dict=boost,
        num_results=10
    )

    return results


evaluate(ground_truth, lambda q: minsearh_improved(q['question']))

### minsearch with filter 

{'Recall': 0.8589318600368324,

 'hit_rate': 0.8589318600368324,
 
 'mrr': 0.7564702271332095}


### Baseline Minsearch
{'Recall': 0.8552486187845304,

 'hit_rate': 0.8552486187845304,
 
 'mrr': 0.7135870677307133}
 

### If the travel assistant is Q&A style (users ask short questions, expect the first result to be correct quickly), go with minsearch with filter → best MRR.

## RAG Flow 

minsearch_search_filter("Can I find halal food?", 'London')


def build_prompt(query, search_results):
    prompt_template = """
You're a travel assistant. Answer the QUESTION based on the CONTEXT from our travel database.
Use only the facts from the CONTEXT when answering the QUESTION.

QUESTION: {question}

CONTEXT:
{context}
""".strip()

    entry_template = """
City: {city}
Section: {section}
Subsection: {subsection}
Text: {text}
""".strip()

    context = ""
    for doc in search_results:
        context += entry_template.format(**doc) + "\n\n"

    prompt = prompt_template.format(question=query, context=context.strip())
    return prompt


from mistralai import Mistral
from mistralai.models import UserMessage
import os
from dotenv import load_dotenv

# loads variables from .env
load_dotenv()  

api_key = os.getenv("API_KEY")

llm_client = Mistral(api_key = api_key)

def llm(prompt):
    response = llm_client.chat.complete(
        model= "ministral-8b-latest",
        messages=[UserMessage(content=prompt)],
    )

    return response.choices[0].message.content

def rag(query, city):
    search_results = minsearch_search_filter(query=query, city=city)
    prompt = build_prompt(query, search_results)
    answer = llm(prompt)

    return answer

answer = rag("Plan 3-day itinerary for Rome?", 'Roma')
print(answer)

### Use minsearch with best paramters

def rag_improved(query):
    search_results = minsearh_improved(query=query)
    prompt = build_prompt(query, search_results)
    answer = llm(prompt)

    return answer

answer = rag_improved("How I can Visit pyrimads in Cairo?")
print(answer)
