==> notebooks/4. Elasticsearch Text Evaluate.ipynb <==
# import os
# os.environ["SSL_CERT_FILE"] = "/mnt/d/Travel Assistant/Musafir/Fortinet_CA_SSL(15).cer"
# os.environ["REQUESTS_CA_BUNDLE"] = "/mnt/d/Travel Assistant/Musafir/Fortinet_CA_SSL(15).cer"

import json
import pandas as pd
from tqdm.auto import tqdm
from sentence_transformers import SentenceTransformer
import os


with open('../data/processed_data/documents-with-ids.json', 'rt') as f_in:
    documents = json.load(f_in)


documents[365]

# Ground Truth data
df_gt = pd.read_csv('../data/result/groud-truth-retrieval.csv')

df_gt

ground_truth = df_gt.to_dict(orient='records')

from elasticsearch import Elasticsearch
import requests


es_client = Elasticsearch('http://localhost:9200')
print(es_client.info())

index_name = "travel_assistant"

if es_client.indices.exists(index=index_name):
    es_client.indices.delete(index=index_name)
    print(f"Delete existing index: {index_name}")

list(documents[0].keys())

# Index doc by 'city' and 'section' value and the rest is the text 

index_settings = {
    "settings":{
        "number_of_shards":1,
        "number_of_replicas":0
    },
    "mappings":{
        "properties":{
            "city":{"type":"keyword"},
            "section":{"type":"text"},
            "subsection":{"type":"text"},
            "text":{"type":"text"},
        }
    }
}

es_client.indices.create(index=index_name, body=index_settings)

# replace none value with null
import math

def clean_doc(doc):
    for k, v in doc.items():
        if isinstance(v, float) and math.isnan(v):
            doc[k] = None  
    return doc

# Set the index for the whole docs

for doc in tqdm(documents):
    es_client.index(index=index_name, document=clean_doc(doc))


## Start the evaluation process

#Hit Rate (HR) or Recall at k
def hit_rate(relevance_total):
    cnt = 0

    for line in relevance_total:
        if True in line:
            cnt = cnt + 1

    return cnt / len(relevance_total)

# Mean Reciprocal Rank (MRR)
def mrr(relevance_total):
    total_score = 0.0

    for line in relevance_total:
        for rank in range(len(line)):
            if line[rank] == True:
                total_score = total_score + 1 / (rank + 1)

    return total_score / len(relevance_total)

def evaluate(ground_truth, search_function):
    relevance_total = []

    for q in tqdm(ground_truth):
        doc_id = q['id']
        results = search_function(q)
        relevance = [d['id'] == doc_id for d in results]
        relevance_total.append(relevance)

    return {
        'hit_rate': hit_rate(relevance_total),
        'mrr': mrr(relevance_total),
    }


## Start testing different elastic search method

# Define the basline elastic search 

def elastic_search(query):
    search_query = {
        "query": {
            "bool": {
                "must": [
                    {
                        "multi_match": {
                            "query": query, 
                            "fields": ['city', 'section', 'subsection', 'text'],
                            "type": "best_fields"
                        }
                    }
                ]
            }
        }
    }

    response = es_client.search(index=index_name, body=search_query)
    result_docs = []

    for hit in response['hits']['hits']:
        result_docs.append(hit['_source'])

    return result_docs
    

Basline_es = evaluate(ground_truth, lambda q: elastic_search(q['question']))
Basline_es

def elastic_search_filter(query, city_filter=None):

    search_query = {
        "size": 5,
        "query": {
            "bool": {
                "must": [
                    {
                        "multi_match": {
                            "query": query,
                            "fields": ['city', 'section', 'subsection', 'text'],
                            "type": "best_fields"
                        }
                    }
                ]
            }
        }
    }

    # Add filter 
    if city_filter:
        search_query["query"]["bool"]["filter"] = {
            "term": {
                "city": city_filter
            }
        }


    response = es_client.search(index=index_name, body=search_query)
    
    result_docs = []

    for hit in response['hits']['hits']:
        result_docs.append(hit['_source'])     
    
    return result_docs

filter_es = evaluate(ground_truth, lambda q: elastic_search_filter(q['question'], q['city']))
filter_es

Basline elastic search 

{'Recall': 0.8629834254143647,

 'hit_rate': 0.8629834254143647,
 
 'mrr': 0.7413546142827905}

### Semantic Search 

### Pre-trained models 
https://www.sbert.net/docs/sentence_transformer/pretrained_models.html

model_name = "multi-qa-distilbert-cos-v1"
model = SentenceTransformer(model_name)

model

for doc in tqdm(documents):
    city = doc['city']
    section = doc['section']
    text = doc['text']
    cst = city + ' ' + section + ' ' + text 

    doc['city_vector'] = model.encode(city)
    doc['section_vector'] = model.encode(section)
    doc['text_vector'] = model.encode(text)
    doc['all_data_vector'] = model.encode(cst)

index_name_vec = "traveller_vector"

index_settings_vect = {
    "settings": {
        "number_of_shards": 1,
        "number_of_replicas": 0
    },
    "mappings": {
        "properties": {
            "city": {"type": "keyword"},
            "section": {"type": "text"},
            "subsection": {"type": "text"},
            "text": {"type": "text"},
            "id": {"type": "keyword"},
            "city_vector":{
                "type": "dense_vector",
                "dims": 768,
                "index": True,
                "similarity": "cosine"
            },
            "section_vector": {
                "type": "dense_vector",
                "dims": 768,
                "index": True,
                "similarity": "cosine"
            }, 
            "text_vector": {
                "type": "dense_vector",
                "dims": 768,
                "index": True,
                "similarity": "cosine"

            },
            "all_data_vector": {
                "type": "dense_vector",
                "dims": 768,
                "index": True,
                "similarity": "cosine"
            }
            
        }
    }
}

es_client.indices.delete(index=index_name_vec, ignore_unavailable=True)
es_client.indices.create(index=index_name_vec, body=index_settings_vect)

for doc in tqdm(documents):
    es_client.index(index=index_name_vec, document=doc)



city = "Cairo"
query = "what transport I should use?"

v_q = model.encode(query)

# Vector (kNN) query
knn_query = {
    "field": "text_vector",
    "query_vector": v_q,
    "k": 5,
    "num_candidates": 10000,
    "boost": 0.5,
    "filter": {
        "term": {
            "city": city
        }
    }
}

# Keyword query
keyword_query = {
    "bool": {
        "must": {
            "multi_match": {
                "query": query,
                "fields": ["city^3", 'section', 'subsection', 'text'], 
                "type": "best_fields",
                "boost": 0.5,
            }
        },
        "filter": {
            "term": {
                "city": city
            }
        }
    }
}

response = es_client.search(
    index=index_name_vec,
    query=keyword_query,
    knn=knn_query,
    size=5
)

### Hybrid search pipeline


def elastic_search_hybrid(field, query, vector, city):
    knn_query = {
        "field": field,
        "query_vector": vector,
        "k": 5,
        "num_candidates": 10000,
        "boost": 0.5,
        "filter": {
            "term": {
                "city": city
            }
        }
    }

    keyword_query = {
        "bool": {
            "must": {
                "multi_match": {
                    "query": query,
                    "fields": ["city^3", 'section', 'subsection', 'text'],
                    "type": "best_fields",
                    "boost": 0.5,
                }
            },
            "filter": {
                "term": {
                    "city": city
                }
            }
        }
    }

    search_query = {
        "knn": knn_query,
        "query": keyword_query,
        "size": 5,
        "_source": ["city", 'section', 'subsection', 'text', "id"]
    }

    es_results = es_client.search(
        index=index_name_vec,
        body=search_query
    )
    
    result_docs = []
    
    for hit in es_results['hits']['hits']:
        result_docs.append(hit['_source'])

    return result_docs

def text_hybrid(q):
    question = q['question']
    city = q['city']

    v_q = model.encode(question)

    return elastic_search_hybrid('text_vector', question, v_q, city)


text_hybrid_es = evaluate(ground_truth, text_hybrid)
text_hybrid_es

def city_hybrid(q):
    question = q['question']
    city = q['city']

    v_q = model.encode(question)

    return elastic_search_hybrid('city_vector', question, v_q, city)

city_hybrid_es = evaluate(ground_truth, city_hybrid)
city_hybrid_es

def section_hybrid(q):
    question = q['question']
    city = q['city']

    v_q = model.encode(question)

    return elastic_search_hybrid('section_vector', question, v_q, city)

section_hybrid_es = evaluate(ground_truth, section_hybrid)
section_hybrid_es

def all_data_hybrid(q):
    question = q['question']
    city = q['city']

    v_q = model.encode(question)

    return elastic_search_hybrid('all_data_vector', question, v_q, city)



all_data_es = evaluate(ground_truth, all_data_hybrid)
all_data_es

### Reranking

def compute_rrf(rank, k=60):
    """ Our own implementation of the relevance score """
    return 1 / (k + rank)

def elastic_search_hybrid_rrf(field, query, vector, course, k=60):
    knn_query = {
        "field": field,
        "query_vector": vector,
        "k": 10,
        "num_candidates": 10000,
        "boost": 0.5,
        "filter": {
            "term": {
                "city": city
            }
        }
    }

    keyword_query = {
        "bool": {
            "must": {
                "multi_match": {
                    "query": query,
                    "fields": ["city^3", 'section', 'subsection', 'text'],
                    "type": "best_fields",
                    "boost": 0.5,
                }
            },
            "filter": {
                "term": {
                    "city": city
                }
            }
        }
    }

    knn_results = es_client.search(
        index=index_name_vec, 
        body={
            "knn": knn_query, 
            "size": 10
        }
    )['hits']['hits']
    
    keyword_results = es_client.search(
        index=index_name_vec, 
        body={
            "query": keyword_query, 
            "size": 10
        }
    )['hits']['hits']
    
    rrf_scores = {}
    # Calculate RRF using vector search results
    for rank, hit in enumerate(knn_results):
        doc_id = hit['_id']
        rrf_scores[doc_id] = compute_rrf(rank + 1, k)

    # Adding keyword search result scores
    for rank, hit in enumerate(keyword_results):
        doc_id = hit['_id']
        if doc_id in rrf_scores:
            rrf_scores[doc_id] += compute_rrf(rank + 1, k)
        else:
            rrf_scores[doc_id] = compute_rrf(rank + 1, k)

    # Sort RRF scores in descending order
    reranked_docs = sorted(rrf_scores.items(), key=lambda x: x[1], reverse=True)
    
    # Get top-K documents by the score
    final_results = []
    for doc_id, score in reranked_docs[:5]:
        doc = es_client.get(index=index_name_vec, id=doc_id)
        final_results.append(doc['_source'])
    
    return final_results

def all_data_hybrid_rrf(q):
    question = q['question']
    city = q['city']

    v_q = model.encode(question)

    return elastic_search_hybrid_rrf('all_data_vector', question, v_q, city)



all_data_hybrid_rrf_es = evaluate(ground_truth, all_data_hybrid_rrf)
all_data_hybrid_rrf_es

from tabulate import tabulate

# build table rows
table = [
    ["Baseline", "Basline_es", Basline_es['hit_rate'], Basline_es['mrr']],
    ["Baseline", "Filter_es", filter_es['hit_rate'], filter_es['mrr']],
    ["Semantic Search", "Text_hybrid_es", text_hybrid_es['hit_rate'], text_hybrid_es['mrr']],
    ["Semantic Search", "City_hybrid_es", city_hybrid_es['hit_rate'], city_hybrid_es['mrr']],
    ["Semantic Search", "Section_hybrid_es", section_hybrid_es['hit_rate'], section_hybrid_es['mrr']],
    ["Semantic Search", "All_data_es", all_data_es['hit_rate'], all_data_es['mrr']],
    ["Reranking", "All_data_hybrid_rrf_es", all_data_hybrid_rrf_es['hit_rate'], all_data_hybrid_rrf_es['mrr']],
]

# print table
print(tabulate(table, headers=["Section", "Method", "Hit Rate", "MRR"], tablefmt="github"))

## RAG Flow 

from mistralai import Mistral
from mistralai.models import UserMessage
import os
from dotenv import load_dotenv

# searching
f_query = "Where should I eat?"
city = "Rome"  

# Encode the query
v_q = model.encode(f_query)

# Run hybrid search on all_data_vector
results = elastic_search_hybrid("all_data_vector", f_query, v_q, city)


# Show results
for i, r in enumerate(results, 1):
    print(f"Result {i}:")
    print(f"  City: {r['city']}")
    print(f"  Section: {r['section']}")
    print(f"  Subsection: {r['subsection']}")
    print(f"  Text: {r['text']}")
    print("-" * 50)

# loads variables from .env
load_dotenv()  

api_key = os.getenv("API_KEY")

client = Mistral(api_key = api_key)

def build_prompt(query, search_results):
    context_template = "Q: {question}\n A: {text}"

    context_parts = []
    for source in search_results:
        context_parts.append(context_template.format(question=source.get('question', ''), text=source.get('text', '')))

    context = "\n\n".join(context_parts)

    prompt_template = """
You're a travel assistant. Answer the QUESTION based on the CONTEXT from the FAQ database.
Use only the facts from the CONTEXT when answering the QUESTION.

QUESTION: {question}

CONTEXT:
{context}

    """.strip()
    
    prompt = prompt_template.format(question=query, context=context)

    return prompt


def llm(prompt):
    response = client.chat.complete(
        model= "mistral-medium-latest",
        messages=[UserMessage(content=prompt)],
    )


    return response.choices[0].message.content

def rag(query):
    v_q = model.encode(query)
    search_results = elastic_search_hybrid("all_data_vector", query, v_q, city)
    prompt = build_prompt(query, search_results)
    answer = llm(prompt)

    return answer

answer = rag(f_query)
print("\n Answer:", answer)

answer = rag("Plan two days trip to Roma")
print("\n Answer:", answer)

answer = rag("Plan three-day itinerary for Dubai")
print("\n Answer:", answer)

answer = rag("what trasnportation I should use in Seoul")
print("\n Answer:", answer)
