==> notebooks/5. Qdrant Text Evaluate.ipynb <==
# import os
# os.environ["SSL_CERT_FILE"] = "/mnt/d/Travel Assistant/Musafir/Fortinet_CA_SSL(15).cer"
# os.environ["REQUESTS_CA_BUNDLE"] = "/mnt/d/Travel Assistant/Musafir/Fortinet_CA_SSL(15).cer"

from qdrant_client import QdrantClient, models
import requests
from fastembed import TextEmbedding
import json
import random
import pandas as pd
from tqdm.auto import tqdm
import uuid
import os


# import the data 
with open('../data/processed_data/documents-with-ids.json', 'rt') as f_in:
    documents = json.load(f_in)


documents[11]

# Ground Truth data
df_gt = pd.read_csv('../data/result/groud-truth-retrieval.csv')

df_gt.head()

ground_truth = df_gt.to_dict(orient='records')

ground_truth[55]

#connecting to local Qdrant instance
qdrant_client = QdrantClient(url="http://localhost:6333")

qdrant_client

emb_models = TextEmbedding.list_supported_models()
print(f"There are {len(emb_models)} models supported")
emb_models[0:10]

model_handle = "jinaai/jina-embeddings-v2-small-en"

## Create collection 

# Define the collection name
collection_name = "traveller-rag"
# Delete a connection 
if qdrant_client.collection_exists(collection_name=collection_name):
    qdrant_client.delete_collection(collection_name=collection_name)
    print(f"Delete the currrent collection {collection_name}")


# Create the collection with specified vector parameters
qdrant_client.create_collection(
    collection_name=collection_name,
    vectors_config=models.VectorParams(
        size = 512,
        distance=models.Distance.COSINE
    )
)

## Create, Embed & Insert Points into the Collection


points = []
id = 0

for doc in documents:
    point = models.PointStruct(
        id=id,
        vector=models.Document(text=doc['text'], model=model_handle),
        payload={
            "id": doc['id'],
            "text": doc['text'],
            "city": doc['city'],
            "section": doc['section'],
            "subsection": doc['subsection']

        }
    )

    points.append(point)

    id+=1

qdrant_client.upsert(
    collection_name=collection_name, 
    points=points
)

# Start the evaluation process

#Hit Rate (HR) or Recall at k
def hit_rate(relevance_total):
    cnt = 0

    for line in relevance_total:
        if True in line:
            cnt = cnt + 1

    return cnt / len(relevance_total)

# Mean Reciprocal Rank (MRR)
def mrr(relevance_total):
    total_score = 0.0

    for line in relevance_total:
        for rank in range(len(line)):
            if line[rank] == True:
                total_score = total_score + 1 / (rank + 1)

    return total_score / len(relevance_total)

def evaluate(ground_truth, search_function):
    relevance_total = []

    for q in tqdm(ground_truth):
        gt_id = q['id']
        results = search_function(q)

        points = results.points if hasattr(results, "points") else results

        relevance = [
            point.payload.get('id') == gt_id
            for point in points
            if hasattr(point, "payload")
        ]
        relevance_total.append(relevance)

    return {
        'hit_rate': hit_rate(relevance_total),
        'mrr': mrr(relevance_total),
    }


# Semantic Search

## Basline

def qdrant_search(query, limit=5):

    results = qdrant_client.query_points(
        collection_name=collection_name,
        query=models.Document(
            text=query,
            model=model_handle
        ),
        limit=limit,
        with_payload=True
    
    )

    return results

basline_qd_search = evaluate(ground_truth, lambda q: qdrant_search(q['question']))
print(basline_qd_search)

return more result improve the evaluation matrix 

## Search with filter


qdrant_client.create_payload_index(
    collection_name=collection_name, 
    field_name="city",
    field_schema="keyword"
)

def qd_search_filter(query, city, limit):
    results=qdrant_client.query_points(
        collection_name=collection_name,
        query=models.Document(
            text=query,
            model=model_handle
        ),
        query_filter=models.Filter(
            must=[
                models.FieldCondition(
                    key="city",
                    match=models.MatchValue(value=city)
                )
            ]
        ),
        limit = limit,
        with_payload=True
    )

    return results

filter_qd_search = evaluate(ground_truth, lambda q: qd_search_filter(q['question'], q['city'], 5))
print(filter_qd_search)

results = qd_search_filter("What museum I should visit?", "Rome", 5)

points = [
    {
        "score": point.score,
        "payload": point.payload
    }
    for point in results.points
]

print(json.dumps(points, indent=2))

# Sparse vector search with BM25

collection_name_sparse = "traveller-sparse"

# Create a collection for sparse vector parameters 
qdrant_client.create_collection(
    collection_name=collection_name_sparse,
    sparse_vectors_config={
        "bm25":models.SparseVectorParams(
            modifier=models.Modifier.IDF
        )
    }
)

# Send the points to the collection
qdrant_client.upsert(
    collection_name=collection_name_sparse,
    points=[
        models.PointStruct(
            id=uuid.uuid4().hex,
            vector={
                "bm25":models.Document(
                    text=doc["text"],
                    model="Qdrant/bm25",
                )
            },
            payload={
            "id": doc['id'],
            "text": doc['text'],
            "city": doc['city'],
            "section": doc['section'],
            "subsection": doc['subsection']

            }
        )
        for doc in documents
    ]
)

## Running sparse vector search with BM25

def bm_search(query: str, limit: int =5) -> list[models.ScoredPoint]:
    results = qdrant_client.query_points(
        collection_name=collection_name_sparse,
        query=models.Document(
            text=query,
            model="Qdrant/bm25",
        ),
        using="bm25",
        limit=limit,
        with_payload=True
    )

    return results.points

bm_qd_search = evaluate(ground_truth, lambda q: bm_search(q['question']))
print(bm_qd_search)

results = bm_search("love")
for point in results:
    print(
        f"Score: {point.score}\n"
        f"City: {point.payload.get('city')}\n"
        f"Section: {point.payload.get('section')}\n"
        f"Subsection: {point.payload.get('subsection')}\n"
        f"Text: {point.payload.get('text')}"
    )
    print("=="*50)
    
        


# Multi-step search dense vector search, and rerank  with sparse search

collection_name_dense_sparse="traveller-dense-and-sparse"

# Create the collection with both vector types
qdrant_client.create_collection(
    collection_name=collection_name_dense_sparse,
    vectors_config={
        'jina-small':models.VectorParams(
            size=512,
            distance=models.Distance.COSINE
        ),
    },
    sparse_vectors_config={
        "bm25":models.SparseVectorParams(
            modifier=models.Modifier.IDF
        )
    }
)

qdrant_client.upsert(
    collection_name=collection_name_dense_sparse,
    points=[
        models.PointStruct(
            id=uuid.uuid4().hex,
            vector={
                "jina-small": models.Document(
                    text=doc["text"],
                    model="jinaai/jina-embeddings-v2-small-en",
                ),
                "bm25": models.Document(
                    text=doc["text"], 
                    model="Qdrant/bm25",
                ),
            },
            payload={
            "id": doc['id'],
            "text": doc['text'],
            "city": doc['city'],
            "section": doc['section'],
            "subsection": doc['subsection']}

        )
        for doc in documents
    ]
)


def dense_sparse_search(query: str, limit: int = 5) -> list[models.ScoredPoint]:
    results = qdrant_client.query_points(
        collection_name=collection_name_dense_sparse,
        prefetch=[
            models.Prefetch(
                query=models.Document(
                    text=query,
                    model="jinaai/jina-embeddings-v2-small-en",
                ),
                using="jina-small",
                limit=(10 * limit),
            ),
        ],
        query=models.Document(
            text=query,
            model="Qdrant/bm25", 
        ),
        using="bm25",
        limit=limit,
        with_payload=True,
    )

    return results.points

dense_sparse_search_eva = evaluate(ground_truth, lambda q: dense_sparse_search(q['question']))
print(dense_sparse_search_eva)

results = dense_sparse_search("what is the best pub in Soho?")
for point in results:
    print(
        f"Score: {point.score}\n"
        f"City: {point.payload.get('city')}\n"
        f"Section: {point.payload.get('section')}\n"
        f"Subsection: {point.payload.get('subsection')}\n"
        f"Text: {point.payload.get('text')}"
    )
    print("=="*50)


# Hybrid Search

def rrf_search(query: str, limit: int = 5) -> list[models.ScoredPoint]:
    results = qdrant_client.query_points(
        collection_name=collection_name_dense_sparse,
        prefetch=[
            models.Prefetch(
                query=models.Document(
                    text=query,
                    model="jinaai/jina-embeddings-v2-small-en",
                ),
                using="jina-small",
                limit=(5 * limit),
            ),
            models.Prefetch(
                query=models.Document(
                    text=query,
                    model="Qdrant/bm25",
                ),
                using="bm25",
                limit=(5 * limit),
            ),
        ],
        query=models.FusionQuery(fusion=models.Fusion.RRF),
        with_payload=True,
    )

    return results.points

hybrid_search_eva = evaluate(ground_truth, lambda q: rrf_search(q['question']))
print(hybrid_search_eva)

results = rrf_search("Plan three days trip to Cairo?")
for point in results:
    print(
        f"Score: {point.score}\n"
        f"City: {point.payload.get('city')}\n"
        f"Section: {point.payload.get('section')}\n"
        f"Subsection: {point.payload.get('subsection')}\n"
        f"Text: {point.payload.get('text')}"
    )
    print("=="*50)


# Evaluation Summary 

from tabulate import tabulate

# build table rows
table = [
    ["Semantic Search", "Semantic Search basline", basline_qd_search['hit_rate'], basline_qd_search['mrr']],
    ["Semantic Search", "Semantic Search with filter", filter_qd_search['hit_rate'], filter_qd_search['mrr']],
    ["Sparse vector", "Sparse vector with BM25", bm_qd_search['hit_rate'], bm_qd_search['mrr']],
    ["Multi-step search", "Dense and sparse", dense_sparse_search_eva['hit_rate'], dense_sparse_search_eva['mrr']],
    ["Hybrid Search", "Hybrid Search", hybrid_search_eva['hit_rate'], hybrid_search_eva['mrr']],
]

# print table
print(tabulate(table, headers=["Section", "Method", "Hit Rate", "MRR"], tablefmt="rounded_grid"))

# RAG Flow

from mistralai import Mistral
from mistralai.models import UserMessage
import os
from dotenv import load_dotenv


# loads variables from .env
load_dotenv()  

api_key = os.getenv("API_KEY")

llm_client = Mistral(api_key = api_key)

def build_prompt(query, search_results):
    context_template = "Q: {question}\n A: {text}"

    context_parts = []
    for point in search_results: 
        payload = point.payload
        context_parts.append(
            context_template.format(
                question=query,
                text=payload.get("text", "")
            )
        )

    context = "\n\n".join(context_parts)
    
    prompt_template = """
You're a travel assistant. Answer the QUESTION based on the CONTEXT from the traveller database.
Use only the facts from the CONTEXT when answering the QUESTION. If the QUESTION doesn't have a CONTEXT data just mention
you cannot answer the question.

QUESTION: {question}

CONTEXT:
{context}

    """.strip()
    
    prompt = prompt_template.format(question=query, context=context)
    return prompt


def llm(prompt):
    response = llm_client.chat.complete(
        model= "ministral-8b-latest",
        messages=[UserMessage(content=prompt)],
    )

    return response.choices[0].message.content

def rag(query):
    search_results = rrf_search(query=query)
    prompt = build_prompt(query, search_results)
    answer = llm(prompt)

    return answer

answer = rag("Plan three trip to Cairo?")
print(answer)

answer = rag("Plan three-day itinerary for Brussels?")
print(answer)
